Refactor QR code
