+++ app/src/androidTest/java/com/sdpteam/connectout/QrCode/QRcodeGeneratorTest.java
@@ -1,4 +1,4 @@
-package com.sdpteam.connectout.qr_code;
+package com.sdpteam.connectout.QrCode;
 
 import static org.junit.Assert.assertNotNull;
