+++ app/src/androidTest/java/com/sdpteam/connectout/QrCode/QRcodeModalActivityTest.java
@@ -1,4 +1,4 @@
-package com.sdpteam.connectout.qr_code;
+package com.sdpteam.connectout.QrCode;
 
 import static androidx.test.espresso.Espresso.onView;
 import static androidx.test.espresso.action.ViewActions.click;
