+++ app/src/main/java/com/sdpteam/connectout/QrCode/CaptureActivityContainer.java
@@ -1,0 +1,10 @@
+package com.sdpteam.connectout.QrCode;
+
+import com.journeyapps.barcodescanner.CaptureActivity;
+
+/**
+ * sole purpose of this class is to override the default CaptureActivity
+ * so that we can change the theme/properties of the QR scanner
+ */
+public class CaptureActivityContainer extends CaptureActivity {
+}
