--- app/src/main/java/com/sdpteam/connectout/qr_code/CaptureAct.java
@@ -1,8 +1,0 @@
-package com.sdpteam.connectout.qr_code;
-
-import com.journeyapps.barcodescanner.CaptureActivity;
-
-// sole purpose of this class is to override the default CaptureActivity
-// so that we can change the theme/properties of the QR scanner
-public class CaptureAct extends CaptureActivity {
-}
