+++ app/src/main/res/layout/activity_qrcode.xml
@@ -3,13 +3,12 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".qr_code.QRcodeActivity">
+    tools:context=".QrCode.QRcodeActivity">
 
     <Button
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="@string/scan"
         android:layout_centerInParent="true"
-        android:id="@+id/btn_scan"
-        />
+        android:id="@+id/btn_scan" />
 </RelativeLayout>
