+++ app/src/main/res/layout/activity_qrcode_modal.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    tools:context=".qr_code.QRcodeModalActivity">
+    tools:context=".QrCode.QRcodeModalActivity">
 
     <TextView
         android:id="@+id/modal_title"
