+++ app/src/main/res/layout/activity_qrcode_profile.xml
@@ -3,14 +3,13 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".qr_code.QRcodeProfileActivity">
+    tools:context=".QrCode.QRcodeProfileActivity">
 
     <Button
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="@string/profile_qr_code"
         android:layout_centerInParent="true"
-        android:id="@+id/show_qr_code_btn"
-        />
+        android:id="@+id/show_qr_code_btn" />
 
 </RelativeLayout>
